feat: implement authorization in Cart's API route and Cart's page

+++ src/app/api/v1/cart/[UserAccountID]/[CartID]/route.ts
@@ -1,12 +1,26 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/db";
 import { ErrorMessages } from "@/src/enums/ErrorMessages";
+import { getServerSession } from "next-auth";
+import { nextAuthOptions } from "@/src/lib/auth";
 
 export async function DELETE(
     req: NextRequest,
  { params }: { params: { UserAccountID: string; CartID: string } }
 ) {
     const { UserAccountID, CartID } = params;
+
+    // If accountId doesn't match the token sent's ID
+    const session = await getServerSession(nextAuthOptions);
+    if (session?.user?.id != UserAccountID) {
+        return new Response(
+            JSON.stringify({
+                success: false,
+                message: ErrorMessages.PERMISSION,
+            }),
+            { status: 401 }
+        );
+    }
 
     try {
         const deletedCart = await prisma.cart.deleteMany({
